+++ package.json
@@ -5,8 +5,7 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "start:dev": "nodemon server.js",
+    "start:dev": "nodemon --watch 'src/**' --ext 'ts, json' --ignore 'src/**/*.spec.ts' --exec 'npx ts-node src/server.ts'",
     "start:prod": ""
   },
   "repository": {
@@ -21,7 +20,14 @@
   },
   "homepage": "https://github.com/shlykovaKate/NodeJS-CRUD-api#readme",
   "devDependencies": {
-    "nodemon": "^2.0.16"
+    "@types/uuid": "^8.3.4",
+    "clean-webpack-plugin": "^4.0.0",
+    "nodemon": "^2.0.16",
+    "ts-node": "^10.8.1",
+    "typescript": "^4.7.4",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
     "uuid": "^8.3.2"
